Separate definition of Provider from that of ASPA
This PR update defines "Provider AS" upfront and avoids defining it inside the definition of the ASPA object. I have read the whole draft and made significant editorial changes throughout. I have inserted some comments with [Sriram: ....] for the authors to consider. There are some repetitions of text and inconsistencies of normative language amongst the repetitions.  -- Sriram