Version 12 Update -- comments incorporated
This updated version 12 is possibly ready for upload to the IETF. I'll wait a couple of days to see if anyone has comments.

It includes all of Claudio's comments (set 1 and set 2).  It also incorporates all comments from Russ, Randy, Alexander, Job, and Nan Geng on the version that I circulated on Feb. 16th.