Updated for version-13 (ietf-sidrops-aspa-profile)
Several edits...
Claudio liked Section 4 in the verification draft and thought it could be repeated in the profile draft.
That material firmly belongs in the verification draft as it relates closely to verification.
I thought there should be no repetition but instead, a strong coupling was needed between the two drafts. 
I have done that with additional text in updated Sections 1 and 3 in the profile draft.
Some repetitions of text are fixed (removed from the Security Considerations). The Security Considerations section is simplified. 
Another comment from Claudio about being specific and clear about non-transparent RS in some places is also incorporated.