Updated for Draft version 14
This revision takes into consideration the comments on v-13 on the sidrops list from Claudio, Igor, and Dai Zhibin. 
Igor read the entire draft (pre-release v-14) and offered many editorial comments and also offered refinements to the Properties in Sec. 8.
I have carefully re-read and edited it for any remaining typos and grammatical errors, etc. 
I would mention to the SIDROPS Chairs that it is ready for the Shepherd report and AD reviews.