Update for v-16 draft of aspa-verification
Some of the authors of the aspa-verification draft (Keyur, Job, Sriram) joined by Ben M. and Oliver B. met in San Francisco (IETF 117) to bring closure on some loose ends in the draft.

The discussion converged to carrying out the following actions with the draft and then uploading v-16:

1. Eliminate Section 7.2  (Verification and Mitigation at Egress eBGP Router).
Reasons: Avoid unnecessary complexity. See slide 8 of the IETF 117 presentation:  
https://datatracker.ietf.org/meeting/117/materials/slides-117-sidrops-aspa-draft-update 
The OTC Attribute (see #3 below) sufficiently helps to prevent route leaks at the local AS. 

2. Do not include ASPA path verification in IBGP either. 
Reasons: Performing ASPA path verification and rejection of routes with Invalid AS path at eBGP ingress is sufficient. The AS is expected to maintain a consistent view of RPKI data across all its border routers. 

3. Add the following subsection 7.2 in Section 7:

7.2.  Only to Customer (OTC) Attribute

While the ASPA-based AS path verification method (Section 7.1)  detects and mitigates route leaks that were created by preceding ASes listed in the AS_PATH, it lacks the ability to prevent route leaks created by the local AS.  The use of the Only to Customer (OTC) Attribute [RFC9234] fills in that gap.  An ASPA-compliant implementation MUST also include the procedures utilizing the OTC Attribute as specified in Section 5 of [RFC9234].